LayoutTests/accessibility/smart-invert-expected.txt

       
       
      
DEFAULTS
PASS: filter for IMG#a is 'none'.
PASS: filter for PICTURE#b is 'none'.
PASS: filter for VIDEO#c is 'none'.
PASS: filter for VIDEO#d is 'none'.
PASS: filter for IMG.no-invert is 'none'.
PASS: filter for PICTURE.no-invert is 'none'.
PASS: filter for VIDEO.no-invert is 'none'.
PASS: filter for VIDEO.no-invert is 'none'.
PASS: filter for IMG.preserve-filter is 'grayscale(1)'.
PASS: filter for PICTURE.preserve-filter is 'grayscale(1)'.
PASS: filter for VIDEO.preserve-filter is 'grayscale(1)'.
PASS: filter for VIDEO.preserve-filter is 'grayscale(1)'.

AFTER INVERT-COLORS IS APPLIED
PASS: filter for IMG#a is 'none'. Change once IMG/PICTURE elements are inverted, too.
PASS: filter for PICTURE#b is 'none'. Change once IMG/PICTURE elements are inverted, too.
PASS: filter for VIDEO#c is 'invert(1)'.
PASS: filter for VIDEO#d is 'invert(1)'.
PASS: filter for IMG.fallback is 'none'.
PASS: filter for IMG.no-invert is 'none'.
PASS: filter for PICTURE.no-invert is 'none'.
PASS: filter for VIDEO.no-invert is 'none'.
PASS: filter for VIDEO.no-invert is 'none'.
PASS: filter for IMG.preserve-filter is 'grayscale(1)'.
PASS: filter for PICTURE.preserve-filter is 'grayscale(1)'.
PASS: filter for VIDEO.preserve-filter is 'grayscale(1)'.
PASS: filter for VIDEO.preserve-filter is 'grayscale(1)'.
